# openapi.yaml
openapi: 3.0.3
info:
  title: Rancher AI Chat API
  description: API for managing chats and messages with tag-based filtering
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://0.0.0.0:8080
    description: Production server
paths:
...
  /llm/{llm_name}/models:
    get:
      summary: Get available models for an LLM provider
      description: Retrieves the list of available models for a specific LLM provider (openai, gemini, bedrock, ollama)
      operationId: get_models
      tags:
        - Configuration
      parameters:
        - name: llm_name
          in: path
          description: The name of the LLM provider
          required: true
          schema:
            type: string
            enum: [openai, gemini, bedrock, ollama]
          example: openai
        - name: url
          in: query
          description: Ollama server URL (required for ollama provider)
          required: false
          schema:
            type: string
          example: http://localhost:11434
        - name: region
          in: query
          description: AWS region (required for bedrock provider, e.g. us-east-1, us-west-2)
          required: false
          schema:
            type: string
          example: us-east-1
        - name: bearerToken
          in: query
          description: AWS bearer token (required for bedrock provider)
          required: false
          schema:
            type: string
          example: token_example...
      responses:
        '200':
          description: List of available models
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["gpt-4o", "gpt-4o-mini", "gpt-4", "gpt-3.5-turbo"]
        '400':
          description: Invalid provider or missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Failed to fetch models from provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /settings:
    get:
      summary: Get current agent settings
      description: Retrieves the current configuration settings for the AI agent
      operationId: get_settings
      tags:
        - Configuration
      responses:
        '200':
          description: Current agent settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  storageType:
                    type: string
                    description: Type of storage backend being used
                    example: in-memory
                required:
                  - storageType
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update agent settings
      description: Updates the agent settings stored in the llm-config Kubernetes secret
      operationId: update_settings
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Updated settings (secret data)
                additionalProperties:
                  type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - user does not have permission to update settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
          description: Error message
      required:
        - message
      example:
        message: userId not found
...
    SettingsUpdate:
      type: object
      description: Configuration settings for the AI agent
      properties:
        OPENAI_API_KEY:
          type: string
          description: OpenAI API key
        OPENAI_URL:
          type: string
          description: OpenAI API base URL
        SYSTEM_PROMPT:
          type: string
          description: System prompt for the agent
        OLLAMA_MODEL:
          type: string
          description: Model to use with Ollama
        GEMINI_MODEL:
          type: string
          description: Model to use with Gemini
        OPENAI_MODEL:
          type: string
          description: Model to use with OpenAI
        BEDROCK_MODEL:
          type: string
          description: Model to use with Bedrock
        GENERIC_OPENAI_MODEL:
          type: string
          description: Model to use with a generic OpenAI-compatible API
        OLLAMA_URL:
          type: string
          description: URL to Ollama server
        GOOGLE_API_KEY:
          type: string
          description: Google API key for Gemini
        ACTIVE_LLM:
          type: string
          enum: [openai, gemini, bedrock, ollama]
          description: Currently active LLM provider
        ENABLE_RAG:
          type: string
          description: Enable RAG (true/false)
        EMBEDDINGS_MODEL:
          type: string
          description: Model to use for embeddings
        LANGFUSE_HOST:
          type: string
          description: Langfuse server host
        LANGFUSE_PUBLIC_KEY:
          type: string
          description: Langfuse public key
        LANGFUSE_SECRET_KEY:
          type: string
          description: Langfuse secret key
        AWS_REGION:
          type: string
          description: AWS region for Bedrock
        AWS_BEARER_TOKEN_BEDROCK:
          type: string
          description: AWS bearer token for Bedrock
        GENERIC_OPENAI_URL:
          type: string
          description: Base URL for a generic OpenAI-compatible provider
        GENERIC_OPENAI_API_KEY:
          type: string
          description: API key for a generic OpenAI-compatible provider
      example:
        ACTIVE_LLM: openai
        OPENAI_API_KEY: sk-...
        OPENAI_MODEL: gpt-4o
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: R_SESS
      description: Rancher session cookie for authentication
security:
  - cookieAuth: []
...